+++ packages/theme/src/context.ts
@@ -4,6 +4,7 @@ import type { ThemeProviderSettings } from './types';
 interface ThemeContextType {
 	resolvedSettings: ThemeProviderSettings;
 	hasCustomPrimary: boolean;
+	hasCustomBackground: boolean;
 }
 
 export const ThemeContext = createContext< ThemeContextType >( {
@@ -13,4 +14,5 @@ export const ThemeContext = createContext< ThemeContextType >( {
 		cornerRadius: undefined,
 	},
 	hasCustomPrimary: false,
+	hasCustomBackground: false,
 } );